pyproject.toml
[tool.poetry]
name = "wikipedia-automation-test"
version = "0.1.0"
description = ""
authors = ["Silver <testingbugsrus@gmail.com>"]
readme = "README.md"
packages = [{include = "wikipedia_automation_test"}]

[tool.poetry.dependencies]
python = "^3.11"
pytest = "7.4.0"
selene = "2.0.0b17"
python-dotenv = "1.0.0"
pydantic = "1.10.9"
requests = "2.31.0"
curlify = "2.2.1"
jsonschema = "4.17.3"
beautifulsoup4 = "4.12.0"
allure-pytest = "2.13.2"
selenium = "4.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
